floating point error solved

+++ TACHE8/test_simplification_contour_bezier2.c
@@ -4,7 +4,7 @@
 int main(int argc, char** argv)
 {
 
-    int d;
+    double d;
     Image I;
     Contour C;
     Robot R;
@@ -15,7 +15,7 @@ int main(int argc, char** argv)
 	
     printf("\n\nFichier : %s\n",argv[1]);
 
-    sscanf(argv[2], "%d", &d);
+    sscanf(argv[2], "%lf", &d);
 
     Image M = mask_detection(I);
     
@@ -66,5 +66,5 @@ int main(int argc, char** argv)
 
     printf("Nombre de Courbes : %d\n", nombre_courbes);
     printf("Nombre de Contours : %d\n", nombre_contours);
-    printf("d = %d\n", d);
+    printf("d = %lf\n", d);
 }
